functions/requirements.txt
-i https://pypi.org/simple
cachecontrol==0.12.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==7.0
firebase-admin==2.13.0
flask==1.0.2
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
msgpack==0.5.6
numpy==1.15.2
pandas==0.23.4
protobuf==3.6.1
pyasn1-modules==0.2.2
pyasn1==0.4.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==4.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
urllib3==1.23
werkzeug==0.14.1
